refactor(dto): Move user-related DTOs to user subpackage
