+++ src/main/java/com/niceone/sharekit/controller/AuthController.java
@@ -3,9 +3,9 @@ package com.niceone.sharekit.controller;
 import com.google.firebase.auth.FirebaseAuthException;
 import com.google.firebase.auth.FirebaseToken;
 import com.niceone.sharekit.domain.user.User;
-import com.niceone.sharekit.dto.LocalLoginRequest;
-import com.niceone.sharekit.dto.SignupRequest;
-import com.niceone.sharekit.dto.UserResponse;
+import com.niceone.sharekit.dto.user.LocalLoginRequest;
+import com.niceone.sharekit.dto.user.SignupRequest;
+import com.niceone.sharekit.dto.user.UserResponse;
 import com.niceone.sharekit.service.FirebaseService;
 import com.niceone.sharekit.service.JwtTokenProvider;
 import com.niceone.sharekit.service.UserService;
