+++ src/main/java/com/niceone/sharekit/dto/user/UserResponse.java
@@ -1,4 +1,4 @@
-package com.niceone.sharekit.dto;
+package com.niceone.sharekit.dto.user;
 
 import com.niceone.sharekit.domain.user.User;
 import lombok.Getter;
